refactor: CommentService 인터페이스 제거
